+++ lib/pay/paddle/webhooks/subscription_payment_refunded.rb
@@ -2,11 +2,11 @@ module Pay
   module Paddle
     module Webhooks
       class SubscriptionPaymentRefunded
-        def initialize(data)
-          charge = Pay.charge_model.find_by(processor: :paddle, processor_id: data["subscription_payment_id"])
+        def call(event)
+          charge = Pay.charge_model.find_by(processor: :paddle, processor_id: event["subscription_payment_id"])
           return unless charge.present?
 
-          charge.update(amount_refunded: Integer(data["gross_refund"].to_f * 100))
+          charge.update(amount_refunded: Integer(event["gross_refund"].to_f * 100))
           notify_user(charge.owner, charge)
         end