+++ lib/pay/paddle/webhooks/subscription_updated.rb
@@ -2,29 +2,29 @@ module Pay
   module Paddle
     module Webhooks
       class SubscriptionUpdated
-        def initialize(data)
-          subscription = Pay.subscription_model.find_by(processor: :paddle, processor_id: data["subscription_id"])
+        def call(event)
+          subscription = Pay.subscription_model.find_by(processor: :paddle, processor_id: event["subscription_id"])
 
           return if subscription.nil?
 
-          case data["status"]
+          case event["status"]
           when "deleted"
             subscription.status = "canceled"
-            subscription.ends_at = Time.zone.parse(data["next_bill_date"]) || Time.zone.now if subscription.ends_at.blank?
+            subscription.ends_at = Time.zone.parse(event["next_bill_date"]) || Time.zone.now if subscription.ends_at.blank?
           when "trialing"
             subscription.status = "trialing"
-            subscription.trial_ends_at = Time.zone.parse(data["next_bill_date"])
+            subscription.trial_ends_at = Time.zone.parse(event["next_bill_date"])
           when "active"
             subscription.status = "active"
-            subscription.paddle_paused_from = Time.zone.parse(data["paused_from"]) if data["paused_from"].present?
+            subscription.paddle_paused_from = Time.zone.parse(event["paused_from"]) if event["paused_from"].present?
           else
-            subscription.status = data["status"]
+            subscription.status = event["status"]
           end
 
-          subscription.quantity = data["new_quantity"]
-          subscription.processor_plan = data["subscription_plan_id"]
-          subscription.paddle_update_url = data["update_url"]
-          subscription.paddle_cancel_url = data["cancel_url"]
+          subscription.quantity = event["new_quantity"]
+          subscription.processor_plan = event["subscription_plan_id"]
+          subscription.paddle_update_url = event["update_url"]
+          subscription.paddle_cancel_url = event["cancel_url"]
 
           # If user was on trial, their subscription ends at the end of the trial
           subscription.ends_at = subscription.trial_ends_at if subscription.on_trial?