+++ lib/pay/stripe.rb
@@ -13,7 +13,6 @@ module Pay
     def setup
       ::Stripe.api_key = private_key
       ::Stripe.api_version = "2020-08-27"
-      ::StripeEvent.signing_secret = signing_secret
 
       # Used by Stripe to identify Pay for support
       ::Stripe.set_app_info("PayRails", partner_id: "pp_partner_IqhY0UExnJYLxg", version: Pay::VERSION, url: "https://github.com/pay-rails/pay")