+++ test/pay/paddle/webhooks/subscription_updated_test.rb
@@ -10,14 +10,14 @@ class Pay::Paddle::Webhooks::SubscriptionUpdatedTest < ActiveSupport::TestCase
     @user.subscriptions.create!(processor: :paddle, processor_id: "does-not-exist", name: "default", processor_plan: "some-plan", status: "active")
 
     Pay.subscription_model.any_instance.expects(:save).never
-    Pay::Paddle::Webhooks::SubscriptionUpdated.new(@data)
+    Pay::Paddle::Webhooks::SubscriptionUpdated.new.call(@data)
   end
 
   test "subscription is updated" do
     @user = User.create!(email: "gob@bluth.com", processor: :paddle, processor_id: @data["user_id"])
     subscription = @user.subscriptions.create!(processor: :paddle, processor_id: @data["subscription_id"], name: "default", processor_plan: "some-plan", status: "active")
 
-    Pay::Paddle::Webhooks::SubscriptionUpdated.new(@data)
+    Pay::Paddle::Webhooks::SubscriptionUpdated.new.call(@data)
     subscription.reload
 
     assert_equal 2, subscription.quantity
@@ -34,7 +34,7 @@ class Pay::Paddle::Webhooks::SubscriptionUpdatedTest < ActiveSupport::TestCase
 
     @data["status"] = "trialing"
 
-    Pay::Paddle::Webhooks::SubscriptionUpdated.new(@data)
+    Pay::Paddle::Webhooks::SubscriptionUpdated.new.call(@data)
 
     assert_equal Time.zone.parse(@data["next_bill_date"]), subscription.reload.trial_ends_at
     assert_nil subscription.reload.ends_at
@@ -46,7 +46,7 @@ class Pay::Paddle::Webhooks::SubscriptionUpdatedTest < ActiveSupport::TestCase
 
     @data["status"] = "deleted"
 
-    Pay::Paddle::Webhooks::SubscriptionUpdated.new(@data)
+    Pay::Paddle::Webhooks::SubscriptionUpdated.new.call(@data)
 
     assert_equal Time.zone.parse(@data["next_bill_date"]), subscription.reload.ends_at
   end
@@ -59,7 +59,7 @@ class Pay::Paddle::Webhooks::SubscriptionUpdatedTest < ActiveSupport::TestCase
     Pay.subscription_model.any_instance.stubs(:on_trial?).returns(true)
     Pay.subscription_model.any_instance.stubs(:trial_ends_at).returns(3.days.from_now.beginning_of_day)
 
-    Pay::Paddle::Webhooks::SubscriptionUpdated.new(@data)
+    Pay::Paddle::Webhooks::SubscriptionUpdated.new.call(@data)
 
     assert_equal 3.days.from_now.beginning_of_day, subscription.reload.ends_at
   end