+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Unreleased
 
+* [BREAKING] Webhooks that can't be verified respond with 400 instead of 200 - @excid3
+* [BREAKING] Remove StripeEvent dependency - excid3
+* [NEW] Add `Pay::Webhook.delegator` for subscribing to webhooks - @excid3
+
 ### 2.4.4
 
 * [Fix] Fixed missing require for version file - @excid3